Lecture_2/Lecture2_Homework/data_format_translation_servers/python_server/pyproject.toml
[tool.poetry]
name = "python_server"
version = "0.1.0"
description = ""
authors = ["Alex <a.sandrovschii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.82.0"
pandas = "^1.4.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
